+++ src/utils/mapUtils.ts
@@ -7,6 +7,7 @@ import bbox from '@turf/bbox';
 import { CollaborativesearchService } from 'arlas-web-core/services/collaborativesearch.service';
 import { map } from 'rxjs/internal/operators/map';
 import * as meta from '@turf/meta';
+import { bboxes } from 'ngeohash';
 
 export function getBounds(elementidentifier: ElementIdentifier, collaborativeSearcheService: CollaborativesearchService)
     : Observable<Array<Array<number>>> {
@@ -33,6 +34,47 @@ export function getBounds(elementidentifier: ElementIdentifier, collaborativeSea
                 const maxY = box[3];
                 return [[minX, minY], [maxX, maxY]];
             }));
+}
+
+export function extentToGeohashes(extent: Array<number>, zoom: number,
+    granularityFunction: (zoom: number) => {tilesPrecision: number, requestsPrecision: number}): Set<string> {
+    let geohashList = [];
+    const west = extent[1];
+    const east = extent[3];
+    const south = extent[2];
+    const north = extent[0];
+    if (west < -180 && east > 180) {
+      geohashList = bboxes(Math.min(south, north),
+        -180,
+        Math.max(south, north),
+        180, Math.max(granularityFunction(zoom).tilesPrecision, 1));
+    } else if (west < -180 && east < 180) {
+      const geohashList_1: Array<string> = bboxes(Math.min(south, north),
+        Math.min(-180, west + 360),
+        Math.max(south, north),
+        Math.max(-180, west + 360), Math.max(granularityFunction(zoom).tilesPrecision, 1));
+      const geohashList_2: Array<string> = bboxes(Math.min(south, north),
+        Math.min(east, 180),
+        Math.max(south, north),
+        Math.max(east, 180), Math.max(granularityFunction(zoom).tilesPrecision, 1));
+      geohashList = geohashList_1.concat(geohashList_2);
+    } else if (east > 180 && west > -180) {
+      const geohashList_1: Array<string> = bboxes(Math.min(south, north),
+        Math.min(180, east - 360),
+        Math.max(south, north),
+        Math.max(180, east - 360), Math.max(granularityFunction(zoom).tilesPrecision, 1));
+      const geohashList_2: Array<string> = bboxes(Math.min(south, north),
+        Math.min(west, -180),
+        Math.max(south, north),
+        Math.max(west, -180), Math.max(granularityFunction(zoom).tilesPrecision, 1));
+      geohashList = geohashList_1.concat(geohashList_2);
+    } else {
+      geohashList = bboxes(Math.min(south, north),
+        Math.min(east, west),
+        Math.max(south, north),
+        Math.max(east, west), Math.max(granularityFunction(zoom).tilesPrecision, 1));
+    }
+    return new Set(geohashList);
 }
 
 export function tileToString(tile: { x: number, y: number, z: number }): string {
